Crate fatality

Expansion for fatal-annotations

Expand #[fatal] annotations on enum variants into two additional enums that can be converted back, or the original split into two.

Stands on the shoulders of thiserror.

Re-exports

pub use thiserror;

Traits

Fatality

Determine the fatality of an error.

Nested

Converts a flat, yet splitable error into a nested Result<Result<_,Jfyi>, Fatal> error type.

Split

Allows to split an error into two types - a fatal and a informational enum error type, that can be further consumed.

Attribute Macros

fatality